+++ torch2trt/converters/__init__.py
@@ -41,7 +41,6 @@ from .sum import *
 from .view import *
 from .tanh import *
 from .transpose import *
-from .UPSAMPLE import * 
 from .upsample import * 
 from .mean import *
 from .softmax import *
